refactor: separating files at test code
